follow

+++ js/src/ctypes/CTypes.cpp
@@ -1259,8 +1259,8 @@ static bool ArgumentLengthError(JSContext* cx, const char* fun,
   return false;
 }
 
-static bool ArrayLengthMismatch(JSContext* cx, unsigned expectedLength,
-                                HandleObject arrObj, unsigned actualLength,
+static bool ArrayLengthMismatch(JSContext* cx, size_t expectedLength,
+                                HandleObject arrObj, size_t actualLength,
                                 HandleValue actual, ConversionType convType) {
   MOZ_ASSERT(arrObj && CType::IsCType(arrObj));
 
@@ -3642,15 +3642,15 @@ static bool ImplicitConvert(JSContext* cx, HandleValue val,
           // Check that array is consistent with type, then
           // copy the array.
           const bool bufferShared = cls == ESClass::SharedArrayBuffer;
-          uint32_t sourceLength =
-              bufferShared ? JS::GetSharedArrayBufferByteLength(valObj)
-                           : JS::GetArrayBufferByteLength(valObj);
+          size_t sourceLength = bufferShared
+                                    ? JS::GetSharedArrayBufferByteLength(valObj)
+                                    : JS::GetArrayBufferByteLength(valObj);
           size_t elementSize = CType::GetSize(baseType);
           size_t arraySize = elementSize * targetLength;
-          if (arraySize != size_t(sourceLength)) {
+          if (arraySize != sourceLength) {
             MOZ_ASSERT(!funObj);
-            return ArrayLengthMismatch(cx, arraySize, targetType,
-                                       size_t(sourceLength), val, convType);
+            return ArrayLengthMismatch(cx, arraySize, targetType, sourceLength,
+                                       val, convType);
           }
           SharedMem<void*> target = SharedMem<void*>::unshared(buffer);
           JS::AutoCheckCannotGC nogc;
@@ -3673,13 +3673,13 @@ static bool ImplicitConvert(JSContext* cx, HandleValue val,
                              arrObj, arrIndex);
           }
 
-          uint32_t sourceLength = JS_GetTypedArrayByteLength(valObj);
+          size_t sourceLength = JS_GetTypedArrayByteLength(valObj);
           size_t elementSize = CType::GetSize(baseType);
           size_t arraySize = elementSize * targetLength;
-          if (arraySize != size_t(sourceLength)) {
+          if (arraySize != sourceLength) {
             MOZ_ASSERT(!funObj);
-            return ArrayLengthMismatch(cx, arraySize, targetType,
-                                       size_t(sourceLength), val, convType);
+            return ArrayLengthMismatch(cx, arraySize, targetType, sourceLength,
+                                       val, convType);
           }
           SharedMem<void*> target = SharedMem<void*>::unshared(buffer);
           JS::AutoCheckCannotGC nogc;
@@ -7932,7 +7932,7 @@ bool CData::ReadStringReplaceMalformed(JSContext* cx, unsigned argc,
                           js::StringBufferArena);
 }
 
-using TypedArrayConstructor = JSObject* (*)(JSContext*, uint32_t);
+using TypedArrayConstructor = JSObject* (*)(JSContext*, size_t);
 
 template <typename Type>
 TypedArrayConstructor GetTypedArrayConstructorImpl() {
